[refactor] Divide interface DDD 형식으로 분리 related to #36
