+++ search/src/main/java/com/bjcareer/search/application/port/in/RankingUsecase.java
@@ -1,0 +1,8 @@
+package com.bjcareer.search.application.port.in;
+
+import java.util.List;
+
+public interface RankingUsecase {
+	void updateKeyword(String keyword);
+	List<String> getRankKeyword(Integer index);
+}
