+++ search/src/main/java/com/bjcareer/search/application/port/search/RankingService.java
@@ -1,9 +1,10 @@
-package com.bjcareer.search.service;
+package com.bjcareer.search.application.port.search;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.bjcareer.search.application.port.in.RankingUsecase;
 import com.bjcareer.search.repository.cache.CacheRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -12,7 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class RankingService {
+public class RankingService implements RankingUsecase {
 	private final CacheRepository cacheRepository;
 
 	public void updateKeyword(String keyword) {
