+++ search/src/main/java/com/bjcareer/search/in/api/controller/dto/QueryRequestDTO.java
@@ -1,4 +1,4 @@
-package com.bjcareer.search.controller.dto;
+package com.bjcareer.search.in.api.controller.dto;
 
 import lombok.Data;
